@@ -1,33 +1,25 @@
 {
   "devDependencies": {
-    "@bazel/concatjs": "^3.6.0",
-    "@bazel/labs": "^3.6.0",
-    "@bazel/protractor": "^3.6.0",
-    "@bazel/rollup": "^3.6.0",
-    "@bazel/terser": "^3.6.0",
-    "@bazel/typescript": "^3.6.0",
-    "@rollup/plugin-commonjs": "14.0.0",
-    "@rollup/plugin-node-resolve": "8.4.0",
+    "@babel/core": "^7.14.6",
+    "@babel/preset-env": "^7.14.7",
+    "@bazel/labs": "3.6.0",
+    "@bazel/rollup": "3.6.0",
+    "@bazel/terser": "3.6.0",
+    "@bazel/typescript": "3.6.0",
+    "@rollup/plugin-typescript": "^8.2.1",
+    "@types/google-protobuf": "^3.15.2",
     "@types/jasmine": "2.8.2",
+    "@types/jest": "^26.0.23",
     "@types/long": "^4.0.0",
-    "@types/node": "11.11.1",
-    "google-protobuf": "3.11.4",
-    "grpc-web": "1.1.0",
-    "http-server": "^0.11.1",
-    "jasmine": "2.8.0",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-firefox-launcher": "1.1.0",
-    "karma-jasmine": "2.0.1",
-    "karma-requirejs": "1.1.0",
-    "karma-sourcemap-loader": "0.3.7",
-    "long": "4.0.0",
-    "protobufjs": "5.0.3",
-    "protractor": "^5.4.2",
-    "requirejs": "2.3.6",
-    "rollup": "2.3.4",
-    "terser": "4.3.1",
-    "typescript": "^3.3.1"
+    "@types/node": "^15.12.4",
+    "babel-jest": "^27.0.5",
+    "google-protobuf": "3.17.3",
+    "grpc-web": "1.2.1",
+    "jest": "^27.0.5",
+    "jest-cli": "^27.0.5",
+    "rollup": "2.52.2",
+    "terser": "5.7.0",
+    "typescript": "4.3.4"
   },
   "scripts": {
     "test": "bazel test ..."
